@@ -1,6 +1,6 @@
 {
   "name": "pattern-string-material-ui",
-  "version": "3.7.0",
+  "version": "4.0.0",
   "description": "A material-ui component which looks like a TEXTFIELD grouping together a dropdown and an input components",
   "main": "lib/index.js",
   "sideEffects": false,
@@ -40,34 +40,34 @@
   },
   "homepage": "https://github.com/iulian-radu-at/pattern-string-material-ui#readme",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
+    "@mui/material": "^5.10.17",
     "lodash": "^4.17.21",
-    "react-select-material-ui": "^7.1.0"
+    "react-select-material-ui": "^9.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@storybook/react": "6.3.8",
+    "@babel/core": "^7.20.5",
+    "@storybook/builder-webpack5": "^6.5.14",
+    "@storybook/manager-webpack5": "^6.5.14",
+    "@storybook/react": "^6.5.14",
     "@types/babel-core": "^6.25.7",
-    "@types/enzyme": "^3.10.9",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/lodash": "^4.14.172",
-    "@types/react": "^16.14.14",
-    "@types/react-dom": "^16.9.14",
-    "@types/react-select": "^4.0.17",
+    "@types/lodash": "^4.14.191",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-select": "^5.0.1",
     "@types/rimraf": "^3.0.2",
     "babel-core": "^6.26.3",
-    "babel-jest": "^27.1.0",
-    "babel-loader": "^8.2.2",
-    "babel-preset-react-app": "^10.0.0",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
+    "babel-preset-react-app": "^10.0.1",
     "cross-env": "^7.0.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
-    "typescript": "^4.4.2"
+    "typescript": "^4.9.4"
   },
   "peerDependencies": {
-    "react": ">=16.9.0",
-    "react-dom": ">=16.9.0"
+    "react": ">=18.0.0",
+    "react-dom": ">=18.0.0"
   },
   "jest": {
     "modulePathIgnorePatterns": [
